Index: winscard_scf.c =================================================================== RCS file: /cvsroot/pcsclite/PCSC/src/winscard_scf.c,v retrieving revision 1.8 diff -r1.8 winscard_scf.c 52c52 <   BYTE bufferLength; --- >   LONG bufferLength;
The byte buffer length would break any APDU over 255 bytes.


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@567 0ce88b0d-b2fd-0310-8134-9614164e65ea

+++ src/winscard_scf.c
@@ -49,7 +49,7 @@ SCARD_IO_REQUEST g_rgSCardT0Pci, g_rgSCardT1Pci, g_rgSCardRawPci;
 static struct _psTransmitMap {
   BYTE Buffer[266];
   int isResponseCached;
-  BYTE bufferLength;
+  LONG bufferLength;
 } psTransmitMap[PCSCLITE_MAX_APPLICATION_CONTEXTS];
 
 /* Channel Map to manage Card Connections. */
